rename dosapp/registry to dosapp/pacman
