+++ pacman/package.go
@@ -1,4 +1,4 @@
-package registry
+package pacman
 
 import (
 	"archive/tar"
