+++ pacman/registry.go
@@ -1,4 +1,4 @@
-package registry
+package pacman
 
 import (
 	"net/url"
