+++ pacman/stage.go
@@ -1,4 +1,4 @@
-package registry
+package pacman
 
 import (
 	"fmt"
